+++ R/get_genome_annotation.R
@@ -36,12 +36,11 @@
 #' @export
 get_genome_annotation <- function(data_type = c("chr_size", "centro_loc", "cytobands", "transcript"),
                                   chrs = paste0("chr", c(1:22, "X", "Y")),
-                                  genome_build = c("hg19", "hg38", "mm10","mm9")) {
+                                  genome_build = c("hg19", "hg38", "mm10", "mm9")) {
   data_type <- match.arg(data_type)
   genome_build <- match.arg(genome_build)
 
-  prefix <- switch(
-    data_type,
+  prefix <- switch(data_type,
     chr_size = "chromsize",
     centro_loc = "centromeres",
     cytobands = "cytobands",