+++ R/show_cn_group_profile.R
@@ -54,7 +54,7 @@ show_cn_group_profile <- function(data,
                                   fill_area = TRUE,
                                   cols = NULL,
                                   chrs = paste0("chr", c(1:22, "X")),
-                                  genome_build = c("hg19", "hg38", "mm10"),
+                                  genome_build = c("hg19", "hg38", "mm10", "mm9"),
                                   cutoff = 2L,
                                   resolution_factor = 1L,
                                   force_y_limit = TRUE,
@@ -134,11 +134,14 @@ show_cn_group_profile <- function(data,
   ))
 
   if (!is.null(highlight_genes)) {
-    gene_file <- switch(
-      genome_build,
+    gene_file <- switch(genome_build,
       mm10 = file.path(
         system.file("extdata", package = "sigminer"),
         "mouse_mm10_gene_info.rds"
+      ),
+      mm9 = file.path(
+        system.file("extdata", package = "sigminer"),
+        "mouse_mm9_gene_info.rds"
       ),
       file.path(
         system.file("extdata", package = "sigminer"),