+++ R/show_cn_profile.R
@@ -47,7 +47,7 @@ show_cn_profile <- function(data, samples = NULL,
                             show_labels = NULL,
                             chrs = paste0("chr", 1:22),
                             position = NULL,
-                            genome_build = c("hg19", "hg38", "mm10"),
+                            genome_build = c("hg19", "hg38", "mm10", "mm9"),
                             ylim = NULL,
                             nrow = NULL, ncol = NULL,
                             return_plotlist = FALSE) {